refactor(engine): extract guard-stack/user-event init from TradingEngine
strategy.py (the 3538-line live TradingEngine) wired its pre-trade guard stack
(StateCollector/OrderSubmitter/OrderValidator/GuardManager) and Coinbase
user-event handler inline in two large __init__ helper methods. Extract them into
engine_initialization.py as engine-taking functions (3538->3401), following the
engine's existing collaborator-function pattern (telemetry_streaming/_health).

The functions populate the engine instance in place; the populated attributes are
declared in __init__ so typing is preserved. Behavior-preserving construction
wiring; 1225 live_trade tests pass, mypy clean.
